refactor(web): give AppPosition a home in lib/app
It sat in `hooks/`, which the frontend standards keep for hooks with no
feature home. This one has a home: `lib/app` already holds the document
title and the footer content, which are the same kind of thing — what the
application is, and where in it the user stands.

The hook becomes a named export alongside its neighbours, and the three
files that were importing from `lib/app/hooks` twice now do it once.
