+++ web/src/app/nrf/NRFChrome.tsx
@@ -9,9 +9,8 @@ import { Button, LineItemButton, OpenButton, Popover } from "@opal/components";
 import { SvgBubbleText, SvgSearchMenu, SvgSidebar } from "@opal/icons";
 import MinimalMarkdown from "@/components/chat/MinimalMarkdown";
 import { useIsSearchModeAvailable } from "@/lib/settings/hooks";
-import { useCustomFooterContent } from "@/lib/app/hooks";
+import { useAppPosition, useCustomFooterContent } from "@/lib/app/hooks";
 import type { AppMode } from "@/providers/QueryControllerProvider";
-import useAppPosition from "@/hooks/useAppPosition";
 import { useQueryController } from "@/providers/QueryControllerProvider";
 import { useTierAtLeast } from "@/hooks/useTierAtLeast";
 import { Tier } from "@/lib/settings/types";
